backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
boto3==1.29.7
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
nanoid==2.0.0
